requirements.txt
pytesseract==0.3.13
Pillow==10.2.0
tk==0.1.0